resources accessible from other views
