+++ packages/client/hmi-client/src/types/Resource.ts
@@ -1,0 +1,11 @@
+import { Project } from '@/types/Project';
+import { RouteName } from '@/router/routes';
+import { RouteParamsRaw } from 'vue-router';
+
+export type Resource = {
+	route: RouteName;
+	params: RouteParamsRaw;
+	name: string;
+	icon: any;
+	projectAsset: Project;
+};
